# src/main/resources/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Airline Ticket Finder App",
    "description": "The purpose of this app is to help us get the best airline ticket deals available. You have the possibility of choosing your location, destination and the date.",
    "termsOfService": "urn:tos",
    "contact": {},
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Inferred Url"
    }
  ],
...
  "paths": {
    "/api/v1/airlines": {
      "get": {
        "tags": [
          "airline-controller"
        ],
        "summary": "gets all the airlines",
        "operationId": "allAirlineUsingGET",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AirlineModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "post": {
        "tags": [
          "airline-controller"
        ],
        "summary": "adds an airline",
        "operationId": "addAirlineUsingPOST",
        "parameters": [
          {
            "name": "airplaneId",
            "in": "query",
            "required": false,
            "style": "pipeDelimited",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "id",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "style": "form",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AirlineModel": {
        "title": "AirlineModel",
        "type": "object",
        "properties": {
          "airplaneId": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}